% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_exemple.R
\docType{data}
\name{Data_exemple}
\alias{Data_exemple}
\title{Data_exemple}
\format{
A data frame with 2076 rows and 5 variables:
\describe{
\item{ID}{the Id of each subject}
\item{visit}{the time of measurement}
\item{y}{the value of the marker}
\item{time}{the time of the first event observed}
\item{event}{the first event observed : 0 = censoring, 1 = first event, 2 = second event}
}
}
\usage{
Data_exemple
}
\description{
A simulated dataset for the exemple of the FlexVar_JM function.
}
\keyword{datasets}
