% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarJM-package.R
\docType{package}
\name{VarJM-package}
\alias{VarJM}
\alias{VarJM-package}
\title{VarJM: Estimate Joint Models with subject-specific variance}
\description{
This package allows to estimate joint models with a subject-specific variance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LeonieCourcoul/VarJM}
  \item Report bugs at \url{https://github.com/LeonieCourcoul/VarJM/issues}
}

}
\author{
\strong{Maintainer}: Léonie Courcoul \email{leonie.courcoul@u-bordeaux.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
